requirements.txt
python-telegram-bot==13.15
python-docx==0.8.11
python-dotenv==1.0.0
urllib3
